backend/requirements.txt
fastapi
uvicorn
motor
beanie
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
scikit-learn
joblib
numpy
imblearn
tqdm
pandas
openai
httpx